Added streaminfo message and system info module
