+++ src/modules/emoteRegistry.ts
@@ -1,7 +1,6 @@
 import {client} from "../index";
 import FileManager from "./storage";
-import {EmoteRegistryDump, Config} from "../structures";
-import {TextChannel} from "discord.js";
+import {EmoteRegistryDump} from "../structures";
 
 function updateGlobalEmoteRegistry(): void {
     const data: EmoteRegistryDump = {version: 1, list: []};
@@ -39,43 +38,13 @@ client.on("emojiUpdate", () => {
     updateGlobalEmoteRegistry();
 });
 
-client.on("guildCreate", (guild) => {
-    console.log(
-        `[GUILD JOIN] ${guild.name} (${guild.id}) added the bot. Owner: ${guild.owner!.user.tag} (${
-            guild.owner!.user.id
-        }). Updated emote registry.`
-    );
-
-    if (Config.systemLogsChannel) {
-        const channel = client.channels.cache.get(Config.systemLogsChannel);
-
-        if (channel && channel.type === "text") {
-            (channel as TextChannel).send(
-                `TravBot joined: \`${guild.name}\`. The owner of this guild is: \`${guild.owner!.user.tag}\` (\`${
-                    guild.owner!.user.id
-                }\`)`
-            );
-        } else {
-            console.warn(`${Config.systemLogsChannel} is not a valid text channel for system logs!`);
-        }
-    }
-
+client.on("guildCreate", () => {
+    console.log("Updated emote registry.");
     updateGlobalEmoteRegistry();
 });
 
-client.on("guildDelete", (guild) => {
-    console.log(`[GUILD LEAVE] ${guild.name} (${guild.id}) removed the bot. Updated emote registry.`);
-
-    if (Config.systemLogsChannel) {
-        const channel = client.channels.cache.get(Config.systemLogsChannel);
-
-        if (channel && channel.type === "text") {
-            (channel as TextChannel).send(`\`${guild.name}\` (\`${guild.id}\`) removed the bot.`);
-        } else {
-            console.warn(`${Config.systemLogsChannel} is not a valid text channel for system logs!`);
-        }
-    }
-
+client.on("guildDelete", () => {
+    console.log("Updated emote registry.");
     updateGlobalEmoteRegistry();
 });
